requirements.txt
pytelegrambotapi
psycopg2
psycopg2-binary
flask
SQLAlchemy==1.3.18
Flask-SQLAlchemy
Flask-Migrate
oauth2client
python-dateutil
TBForms
flask-redis